Apply @colin-axner suggestions
