+++ relayer/naive-strategy.go
@@ -1,7 +1,6 @@
 package relayer
 
 import (
-	"encoding/hex"
 	"fmt"
 	"strconv"
 	"strings"
@@ -10,7 +9,6 @@ import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	clienttypes "github.com/cosmos/cosmos-sdk/x/ibc/core/02-client/types"
 	chantypes "github.com/cosmos/cosmos-sdk/x/ibc/core/04-channel/types"
-	tmclient "github.com/cosmos/cosmos-sdk/x/ibc/light-clients/07-tendermint/types"
 	ctypes "github.com/tendermint/tendermint/rpc/core/types"
 	"golang.org/x/sync/errgroup"
 )
@@ -20,20 +18,17 @@ var (
 	_ Strategy = &NaiveStrategy{}
 
 	// Strings for parsing events
-	spTag        = "send_packet"
-	waTag        = "write_acknowledgement"
-	srcChanTag   = "packet_src_channel"
-	dstChanTag   = "packet_dst_channel"
-	srcPortTag   = "packet_src_port"
-	dstPortTag   = "packet_dst_port"
-	dataTag      = "packet_data"
-	ackTag       = "packet_ack"
-	toHeightTag  = "packet_timeout_height"
-	toTSTag      = "packet_timeout_timestamp"
-	seqTag       = "packet_sequence"
-	updateCliTag = "update_client"
-	headerTag    = "header"
-	clientIDTag  = "client_id"
+	spTag       = "send_packet"
+	waTag       = "write_acknowledgement"
+	srcChanTag  = "packet_src_channel"
+	dstChanTag  = "packet_dst_channel"
+	srcPortTag  = "packet_src_port"
+	dstPortTag  = "packet_dst_port"
+	dataTag     = "packet_data"
+	ackTag      = "packet_ack"
+	toHeightTag = "packet_timeout_height"
+	toTSTag     = "packet_timeout_timestamp"
+	seqTag      = "packet_sequence"
 )
 
 // NewNaiveStrategy returns the proper config for the NaiveStrategy
@@ -247,56 +242,6 @@ func (nrs *NaiveStrategy) HandleEvents(src, dst *Chain, events map[string][]stri
 	if len(rlyPackets) > 0 && err == nil {
 		nrs.sendTxFromEventPackets(src, dst, rlyPackets)
 	}
-}
-
-func checkAndSubmitMisbehaviour(src *Chain, events map[string][]string) error {
-	if hdrs, ok := events[fmt.Sprintf("%s.%s", updateCliTag, headerTag)]; ok {
-		for i, hdr := range hdrs {
-			clientIDs := events[fmt.Sprintf("%s.%s", updateCliTag, clientIDTag)]
-			emittedClientID := clientIDs[i]
-
-			if src.PathEnd.ClientID == emittedClientID {
-				hdrBytes, err := hex.DecodeString(hdr)
-				if err != nil {
-					return err
-				}
-
-				exportedHeader, err := clienttypes.UnmarshalHeader(src.Encoding.Marshaler, hdrBytes)
-				if err != nil {
-					return err
-				}
-
-				emittedHeader, ok := exportedHeader.(*tmclient.Header)
-				if !ok {
-					return fmt.Errorf("emitted header is not tendermint type")
-				}
-
-				trustedHeader, err := src.GetLightSignedHeaderAtHeight(emittedHeader.Header.Height)
-				if err != nil {
-					return err
-				}
-
-				if !IsMatchingConsensusState(emittedHeader.ConsensusState(), trustedHeader.ConsensusState()) {
-					misbehaviour := tmclient.NewMisbehaviour(emittedClientID, emittedHeader, trustedHeader)
-					msg, err := clienttypes.NewMsgSubmitMisbehaviour(emittedClientID, misbehaviour, src.MustGetAddress())
-					if err != nil {
-						return err
-					}
-					res, success, err := src.SendMsg(msg)
-					if err != nil {
-						return err
-					}
-					if !success {
-						return fmt.Errorf("submit misbehaviour tx failed: %s", res.RawLog)
-					}
-					src.Log(fmt.Sprintf("Submitted misbehaviour for emitted header with height: %d",
-						emittedHeader.Header.Height))
-				}
-			}
-		}
-	}
-
-	return nil
 }
 
 func relayPacketsFromEventListener(src, dst *PathEnd, events map[string][]string) (rlyPkts []relayPacket, err error) {
